Move show descendants button into detail pain
