+++ context/app/static/js/components/Detail/provenance/ProvGraph/ProvGraph.jsx
@@ -1,11 +1,30 @@
-import React from 'react';
+import React, { useContext } from 'react';
 import PropTypes from 'prop-types';
+import Button from '@material-ui/core/Button';
 
 import SectionItem from 'js/components/Detail/SectionItem';
 import { LightBlueLink } from 'js/shared-styles/Links';
+import { AppContext } from 'js/components/Providers';
+import { getImmediateDescendantProv } from 'js/hooks/useImmediateDescendantProv';
+import useProvenanceStore from 'js/stores/useProvenanceStore';
 import ProvVis from '../ProvVis';
 import { StyledTypography, FlexPaper } from './style';
 import '@hms-dbmi-bgm/react-workflow-viz/dist/react-workflow-viz.min.css';
+import ProvData from '../ProvVis/ProvData';
+
+function createStepNameSet(steps) {
+  const nameSet = new Set([]);
+  steps.forEach((step) => nameSet.add(step.name));
+  return nameSet;
+}
+
+function removeExistingSteps(steps, newSteps) {
+  const nameSet = createStepNameSet(steps);
+  const uniqueNewSteps = [...newSteps].filter((step) => !nameSet.has(step.name));
+  return uniqueNewSteps;
+}
+
+const useProvenanceStoreSelector = (state) => ({ steps: state.steps, setSteps: state.setSteps });
 
 function ProvGraph(props) {
   const { provData, display_doi } = props;
@@ -27,39 +46,74 @@ function ProvGraph(props) {
   }
 
   function renderDetailPane(prov) {
-    const typeEl =
-      typeKey in prov ? (
-        <SectionItem label="Type">
-          <StyledTypography variant="body1">{prov[typeKey]}</StyledTypography>
-        </SectionItem>
-      ) : (
-        <SectionItem label="Type">
-          <StyledTypography variant="body1">{prov['prov:type']}</StyledTypography>
-        </SectionItem>
+    function DetailPanel() {
+      const { elasticsearchEndpoint, entityEndpoint, nexusToken } = useContext(AppContext);
+      const { steps, setSteps } = useProvenanceStore(useProvenanceStoreSelector);
+
+      async function getMoreSteps() {
+        const results = await getImmediateDescendantProv(
+          display_doi,
+          elasticsearchEndpoint,
+          entityEndpoint,
+          nexusToken,
+        );
+        const moreSteps = results
+          .map((result) => new ProvData(result, getNameForActivity, getNameForEntity).toCwl())
+          .flat();
+        const uniqueNewSteps = removeExistingSteps(steps, moreSteps);
+
+        const stepsCopy = [...steps];
+        stepsCopy[0].outputs[0].target = [
+          { step: 'Create Sample Activity - HBM229.JTTG.749', name: 'Donor - HBM546.MHVZ.749' },
+        ];
+        const x = [...stepsCopy, ...uniqueNewSteps];
+        setSteps(x);
+      }
+
+      const typeEl =
+        typeKey in prov ? (
+          <SectionItem label="Type">
+            <StyledTypography variant="body1">{prov[typeKey]}</StyledTypography>
+          </SectionItem>
+        ) : (
+          <SectionItem label="Type">
+            <StyledTypography variant="body1">{prov['prov:type']}</StyledTypography>
+          </SectionItem>
+        );
+      const idEl =
+        typeKey in prov && ['Donor', 'Sample', 'Dataset'].includes(prov[typeKey]) ? (
+          <SectionItem label="ID" ml>
+            <StyledTypography variant="body1">
+              <LightBlueLink href={`/browse/${prov[typeKey].toLowerCase()}/${prov['hubmap:uuid']}`}>
+                {prov[idKey]}
+              </LightBlueLink>
+            </StyledTypography>
+          </SectionItem>
+        ) : null;
+      const createdEl =
+        timeKey in prov ? (
+          <SectionItem label="Created" ml>
+            <StyledTypography variant="body1">{prov[timeKey]}</StyledTypography>
+          </SectionItem>
+        ) : null;
+      const actionsEl =
+        typeKey in prov && ['Donor', 'Sample', 'Dataset'].includes(prov[typeKey]) ? (
+          <SectionItem ml>
+            <Button color="primary" variant="contained" onClick={getMoreSteps}>
+              Show Descendants
+            </Button>
+          </SectionItem>
+        ) : null;
+      return (
+        <FlexPaper>
+          {typeEl}
+          {idEl}
+          {createdEl}
+          {actionsEl}
+        </FlexPaper>
       );
-    const idEl =
-      typeKey in prov && ['Donor', 'Sample', 'Dataset'].includes(prov[typeKey]) ? (
-        <SectionItem label="ID" ml>
-          <StyledTypography variant="body1">
-            <LightBlueLink href={`/browse/${prov[typeKey].toLowerCase()}/${prov['hubmap:uuid']}`}>
-              {prov[idKey]}
-            </LightBlueLink>
-          </StyledTypography>
-        </SectionItem>
-      ) : null;
-    const createdEl =
-      timeKey in prov ? (
-        <SectionItem label="Created" ml>
-          <StyledTypography variant="body1">{prov[timeKey]}</StyledTypography>
-        </SectionItem>
-      ) : null;
-    return (
-      <FlexPaper>
-        {typeEl}
-        {idEl}
-        {createdEl}
-      </FlexPaper>
-    );
+    }
+    return <DetailPanel />;
   }
 
   return (
